refactoring comments
