+++ src/routes/accessRoute.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const access = require('../conrollers/accessController');
+const access = require('../controllers/accessController');
 
 /* GET Access role premissions. */
 router.get('/', function(req, res, next) {
